# openapi/v1.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '1.0.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Publishing sends a gzipped tarball of the Twext project (`twext.yml` plus
    sources); the hub validates the manifest and compiles the extension
    itself in a sandboxed build, so the reviewed source is exactly what
    produces the served blob.

    The `/v1` surface documented here is stable as of 1.0.0. Breaking changes
    require a new major version; additive changes bump the minor (see
    CHANGELOG.md).

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they patch their own user with
    `termsAcceptedVersion` set to the `version` that `GET /terms` reports. A
    stale version is a `422`, so an acceptance is always recorded against text
    the account was shown. Exemptions that stay available without
    re-acceptance: `DELETE /sessions/current`, `GET /me`, the acceptance
    itself, password-only user updates, and account deletion.

    Bearer tokens come in two kinds. **Sessions** are created by
...
    every scope. **Automation tokens** are created explicitly via
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /sessions:
    post:
      summary: Sign in
      operationId: createSession
      tags: [Open]
      security: []
      description: >
        Signing in is creating a session, so the credentials ride in the body
        rather than a namespace in the path. The plaintext `token` is shown
        once here and never again: only its hash is stored. A namespace that
        belongs to an organization is a `403` here rather than a bad password:
        an organization has no credentials of its own, and is acted for by its
        owners instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace: { $ref: '#/components/schemas/Namespace' }
                password: { type: string, writeOnly: true }
      responses:
        '201':
          description: The session that was created, its owner, and its one-time token
          headers:
            Location:
              description: The new session's own URL
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [session, user, token]
                properties:
                  session: { $ref: '#/components/schemas/Session' }
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    get:
      summary: List a caller's active sessions
      operationId: listSessions
      tags: [Userspace]
      description: Needs the `manage:sessions` scope. `namespace` defaults to the caller; an admin may pass it to inspect another account's sessions.
      parameters:
        - name: namespace
          in: query
          schema: { $ref: '#/components/schemas/Namespace' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated sessions
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /users:
    post:
      summary: Create an account
      operationId: createUser
      tags: [Open]
      security: []
      description: >
        Creating the account and signing in are the same act, so the response
        carries a session too: the new user, the session opened for it, and the
        plaintext token to use for both. The first account ever created becomes
        an admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace: { $ref: '#/components/schemas/Namespace' }
                displayName: { type: string, maxLength: 80 }
                password: { type: string, minLength: 8, writeOnly: true }
      responses:
        '201':
          description: Account created, with a session already open for it
          headers:
            Location:
              description: The new account's own URL
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [user, session, token]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  session: { $ref: '#/components/schemas/Session' }
                  token: { type: string }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    get:
      summary: List accounts
      operationId: listUsers
      tags: [Open]
      security: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated accounts
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/{namespace}:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: Get an account
      operationId: getUser
      tags: [Open]
      security: []
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update an account
      operationId: updateUser
      tags: [Userspace]
      description: >
        Self or admin. Only an admin may set `role`. Changing `password`
        revokes all of that account's existing tokens, and a non-admin caller
        must also send their `currentPassword`, which is rejected with a `403`
        when it does not match. Profile fields
        (`bio`, `website`, `github`, `avatarUrl`, `bannerUrl`) are
        self-service and do not require re-accepting the Terms.

        An organization is not an account and is a `403` here; its profile is
        edited at `PATCH /orgs/{namespace}`.

        `termsAcceptedVersion` records acceptance of the Terms of Service. It is
        exempt from the gate that every other field is behind, so a caller who
        has not accepted yet can still call this — but only on its own: send it
        alone, or with `password`/`currentPassword`. Sending it next to a
        profile or role edit is a `403`, so acceptance cannot be used to carry
        an edit past the gate. A version other than the one `GET /terms`
        reports is a `422`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                displayName: { type: string, maxLength: 80 }
                password: { type: string, minLength: 8, writeOnly: true }
                currentPassword:
                  type: string
                  minLength: 1
                  writeOnly: true
                  description: >
                    Required when a non-admin caller changes `password`.
                    Ignored otherwise.
                role: { $ref: '#/components/schemas/Role' }
                bio: { type: ['string', 'null'], maxLength: 280 }
                website: { type: ['string', 'null'], maxLength: 400, format: uri }
                github:
                  type: ['string', 'null']
                  description: GitHub username, shown as a profile link.
                avatarUrl:
                  type: ['string', 'null']
                  maxLength: 400
                  format: uri
                  description: >
                    An external avatar to reference, not proxy. This is the
                    secondary option and is kept even when the account has
                    uploaded an avatar, in which case the upload is what
                    `GET /users/{namespace}/avatar` serves and this is the
                    fallback; `DELETE /users/{namespace}/avatar` falls back to
                    it. `null` clears the link and leaves any upload alone.
                bannerUrl:
                  type: ['string', 'null']
                  maxLength: 400
                  format: uri
                  description: >
                    An external banner to reference, not proxy, with the same
                    relationship to an uploaded banner as `avatarUrl` has to an
                    uploaded avatar.
                termsAcceptedVersion:
                  type: ['integer', 'string']
                  description: >
                    The `version` from `GET /terms` that the caller accepts.
                    The users row holds it as a string, so either spelling is
                    compared as text.
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Delete an account
      operationId: deleteUser
      tags: [Userspace]
      description: >
        Self or admin. Cascades to all of that account's extensions and
        versions. A namespace that belongs to an organization is a `403`:
        organizations are removed at `DELETE /orgs/{namespace}`. An account
        that is the last owner of an organization is also a `409`, since the
        cascade would leave the organization unmanageable.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: >
        Opaque position from a previous page, taken from that page's
        `_links.next` or `_links.prev` rather than assembled by hand.
    Dir:
      name: dir
      in: query
      schema: { type: string, enum: [back] }
      description: >
        Set to `back` to page towards the rows before the cursor. The server only
        looks for a cursor from a list sorted the other way round; anything else
        is a `400`. A client never needs it, because every page's `_links.prev`
        already sets it.
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
      description: Values above the server cap are clamped to the cap, not rejected. The default cap is 50.
    NamespacePath:
      name: namespace
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Namespace' }
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
...
    Role:
      type: string
      enum: [admin, normal]
...
    Session:
      type: object
      required: [id, createdAt, expiresAt, lastUsedAt]
      properties:
        id: { type: string, description: "The token row's own id, not a hash of the token." }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, description: '7 days after creation.' }
        lastUsedAt: { type: ['string', 'null'], format: date-time }
...
    Kind:
      type: string
      enum: [user, organization]
      description: >
        Which sort of namespace this is. An organization owns extensions and has
        a profile, but has no password, no sessions and no role, so most account
        endpoints refuse it with a `403` and point at `/orgs/{namespace}`.
    User:
      type: object
      required: [namespace, displayName, role, hasPublished, createdAt]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        kind: { $ref: '#/components/schemas/Kind' }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        hasPublished: { type: boolean }
        bio: { type: string }
        website: { type: ['string', 'null'], format: uri }
        github: { type: ['string', 'null'] }
        avatarUrl:
          type: ['string', 'null']
          format: uri
          description: >
            The external reference, or this account's own
            `/users/{namespace}/avatar` path when the image was uploaded rather
            than referenced. Either way this resolves to the image to display,
            so a consumer does not need to know which it is.
        bannerUrl:
          type: ['string', 'null']
          format: uri
          description: >
            The external reference, or this account's own
            `/users/{namespace}/banner` path after an upload.
        createdAt: { type: string, format: date-time, readOnly: true }
        termsAcceptedVersion:
          type: ['integer', 'null']
          description: Terms of Service `version` this account last accepted, or `null` if never accepted. Present only for self or admin viewers.
      description: >
        An account. `GET /users/{namespace}` also answers for an organization,
        with `kind` set to `organization` and without `role` or
        `termsAcceptedVersion`, since an organization has neither. The
        organization view of itself is `GET /orgs/{namespace}`.
...
    PageLinks:
      type: object
      description: >
        The pages either side of this one, as complete URLs. Each already carries
        the cursor, the limit and the filters it was built from, so a client
        follows a link rather than assembling a query; a null means that
        direction is the end of the list.
      required: [self, next, prev]
      properties:
        self:
          type: string
          description: The request as it was made.
        next: { type: ['string', 'null'] }
        prev: { type: ['string', 'null'] }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Validation Error' }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: >
        Authenticated, but not allowed to perform this action — e.g. not the
        resource owner or an admin, a namespace mismatch on publish, the
        current Terms of Service haven't been accepted yet
        (see `termsAcceptedVersion` on `PATCH /users/{namespace}`), or a
        credential missing the scope the route needs. The `detail` names the
        scope. An automation token on `/admin/*` or a version review also needs
        the `admin` scope on top of the account being an admin.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Resource already exists
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    UnprocessableEntity:
      description: Semantically invalid input (e.g. non-SemVer version, id mismatch)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer